Restructure math toggle package, i.e., move everything to where it belongs
