+++ resources/META-INF/plugin.xml
@@ -572,7 +572,7 @@
         </intentionAction>
 
         <intentionAction>
-            <className>nl.rubensten.texifyidea.intentions.latexmathtoggle.LatexMathToggle</className>
+            <className>nl.rubensten.texifyidea.intentions.LatexMathToggle</className>
             <category>LaTeX</category>
             <descriptionDirectoryName>LatexMathToggle</descriptionDirectoryName>
         </intentionAction>
