+++ test/nl/rubensten/texifyidea/editor/OneLinerTest.kt
@@ -1,21 +1,21 @@
-package nl.rubensten.texifyidea.intentions.latexmathtoggle
+package nl.rubensten.texifyidea.editor
 
 import org.junit.Test
 import kotlin.test.assertEquals
 
-class OneLinerTest {
+class MultiLineMathContentToOneLineTest {
 
     @Test
     fun oneLine() {
         val equation = "f(x, y) = 3x + y"
-        assertEquals(equation, OneLiner(equation).getOneLiner())
+        assertEquals(equation, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
     fun oneLineAlign() {
         val equation = "\\alpha \\beta & asdf \\\\"
         val goal = "\\alpha \\beta  asdf "
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -23,7 +23,7 @@ class OneLinerTest {
         val equation = "f(x, y) &= 3x + y \\\\\n" +
                 "&= 3x + 2y - y"
         val goal = "f(x, y) = 3x + y = 3x + 2y - y"
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -31,7 +31,7 @@ class OneLinerTest {
         val equation = "f(x) = x^{32} + 2 x^{16} + 4 x^8 \\\\\n" +
                 "+ 8 x^4 + 16 x^2 + 32 x"
         val goal = "f(x) = x^{32} + 2 x^{16} + 4 x^8 + 8 x^4 + 16 x^2 + 32 x"
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -40,7 +40,7 @@ class OneLinerTest {
                 "    1 & x < 0 \\\\\n" +
                 "    0 & x \\ge 0\n" +
                 "\\end{cases}"
-        assertEquals(equation, OneLiner(equation).getOneLiner())
+        assertEquals(equation, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -54,7 +54,7 @@ class OneLinerTest {
                 "            1 & x < 0 \\\\\n" +
                 "            0 & x \\ge 0\n" +
                 "\\end{cases} = -_-"
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -74,7 +74,7 @@ class OneLinerTest {
                 "            1 & x < 0 \\\\\n" +
                 "            0 & x \\ge 0\n" +
                 "\\end{cases}"
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -89,7 +89,7 @@ class OneLinerTest {
                 "                   \\end{cases}\n" +
                 "               \\end{cases}\n" +
                 "           \\end{cases}"
-        assertEquals(equation, OneLiner(equation).getOneLiner())
+        assertEquals(equation, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -98,7 +98,7 @@ class OneLinerTest {
                 "    1 & x < 0 \\\\\n" +
                 "    0 & x \\ge 0\n" +
                 "\\end{cases}\\end{split}"
-        assertEquals(equation, OneLiner(equation).getOneLiner())
+        assertEquals(equation, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 
     @Test
@@ -108,6 +108,6 @@ class OneLinerTest {
                 "\\intertext{hello} \n" +
                 "&= y"
         val goal = "x = y = x \\text{hello} = y"
-        assertEquals(goal, OneLiner(equation).getOneLiner())
+        assertEquals(goal, MultiLineMathContentToOneLine(equation).getOneLiner())
     }
 }
